[InstrProf] Fix single byte counters for elided branches
Single byte counters for elided branches in `CGStmt::EmitIfStmt` were not
correctly incremented. In normal mode, the counter bound to the `IfStmt`
counts the `then` arm, but in single-byte mode, it's intended to count the
continuation block.

This patch adds the necessary increments for single-byte coverage mode in the
elided branch case.

Change-Id: I695fd4b46cf5b8e2333cbe820ac98c2e0cf2604d
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>